+++ CHANGELOG.md
@@ -7,8 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Changed
 
-- Renamed `ForceReplace` option to `Overwrite` to better match the
-  behavior of the `kubectl apply` subcommand
+- Renamed `Replace` option to `Overwrite` to better match the behavior
+  of the `kubectl apply` subcommand. To use it: `m.Apply(Overwrite(true))`
   
 ### Added